Update to forge .99 and rebase on 1.18.x changes (#313)
* Update stale.yml (#302)

* update build.gradle and bug-report.yml (#306)

* add api output to jarjar (#308)

* add entity size attribute and shrink effect (#310)

* additional shrink effect stuff

* fix eye height for entities with hardcoded eye height

* Bump to v1.18.2-1.1.1 (#311)

* bump version in bug report template

* add shrink update info to the changelog

* Fix build process (#312)

* Fix build.gradle

* Fix release.yml

* Update changelog.md

* add altar power to TOP info (#314)

* Fix fall damage (#315)

* fix fall damage from low heights

* bump version number, update changelog and run datagen

* port to 1.19 (it compiles)

* fix and run datagen

* it works (except worldgen)

* add worldgen

* merge changes from 1.18.x (#307)

* Update stale.yml (#302)

* update build.gradle and bug-report.yml (#306)

* update changelog.md

* add api output to jarjar (#308)

* bump forge version

* bump jei version

* update to forge .94 (still broken because patchouli hasn't updated yet)

* update to forge .99

* add 1.19 structure materials and fix AltarCoreModel

* fix imports that got messed up by the update script and use StrictNBTIngredient.of instead of ctor

* use helper method to register brewing recipes

* fix release workflow (#316)

* implement requested changes

+++ src/main/java/com/github/minecraftschurlimods/arsmagicalegacy/Config.java
@@ -69,7 +69,7 @@ public final class Config {
         public final ForgeConfigSpec.DoubleValue LEVELING_BASE;
         public final ForgeConfigSpec.DoubleValue LEVELING_MULTIPLIER;
         public final ForgeConfigSpec.IntValue EXTRA_BLUE_SKILL_POINTS;
-        public final ForgeConfigSpec.IntValue DAMAGE;
+        public final ForgeConfigSpec.DoubleValue DAMAGE;
         public final ForgeConfigSpec.IntValue DURATION;
         public final ForgeConfigSpec.DoubleValue DRYAD_BONEMEAL_CHANCE;
         public final ForgeConfigSpec.IntValue DRYAD_BONEMEAL_TIMER;
@@ -138,7 +138,7 @@ public final class Config {
             DAMAGE = builder
                     .comment("Damage of damage-based components, in half hearts.")
                     .translation(TranslationConstants.CONFIG + "spell_parts.damage")
-                    .defineInRange("damage", 6, 1, Short.MAX_VALUE);
+                    .defineInRange("damage", 6.0, 1, Short.MAX_VALUE);
             DURATION = builder
                     .comment("Duration of effect-based components, in ticks.")
                     .translation(TranslationConstants.CONFIG + "spell_parts.duration")
